Add multiple select component to travis-form
