+++ app/components/forms/form-select-multiple.js
@@ -1,0 +1,5 @@
+import EmberPowerSelectMultiple from 'ember-power-select/components/power-select-multiple';
+import FormSelectMixin from 'travis/mixins/components/form-select';
+
+export default EmberPowerSelectMultiple.extend(FormSelectMixin, {
+});
