+++ app/styles/app/modules/travis-form.scss
@@ -1,5 +1,6 @@
 $travis-form-border-width: 4px;
 $travis-form-badge-radius: 9px;
+$select-multiple-spacing: 5px;
 
 @mixin makeFieldBorder($color) {
   box-shadow: inset 0 0 1px 1px $color;
@@ -79,6 +80,10 @@ $travis-form-badge-radius: 9px;
       &:focus,
       &:active {
         @include makeFocused();
+      }
+
+      &.travis-form__field-select--multiple {
+        padding: $select-multiple-spacing 25px 0 $select-multiple-spacing;
       }
 
     }
@@ -337,5 +342,39 @@ $travis-form-badge-radius: 9px;
   &[aria-current="true"] {
     background-color: $oxide-blue;
   }
+}
+
+.ember-power-select-multiple {
+
+  &-option {
+    margin: 0 $select-multiple-spacing $select-multiple-spacing 0;
+    border: none;
+    color: $oxide-blue;
+    background-color: lighten($oxide-blue, 45%);
+    font-size: 1.1em;
+    border-radius: 2px;
+    padding: 3px 6px;
+
+  }
+
+  &-remove-btn {
+    font-size: 1.2em;
+    padding: 0;
+    margin: 0;
+    line-height: 0;
+    vertical-align: middle;
+
+    &:not(:hover) {
+      opacity: 1;
+    }
+
+  }
+
+}
+
+.ember-power-select-trigger-multiple-input {
+  margin: 5px 3px 4px 3px;
+  font-size: 1.1em;
+  max-width: calc(100% - 25px);
 }
