+++ app/templates/components/forms/form-field.hbs
@@ -46,7 +46,7 @@
     onInit=(action 'setFieldElementId')
     class='travis-form__field-component'
   )
-  select=(component 'forms/form-select'
+  select=(component selectComponent
     form=form
     selected=value
     disabled=disabled
